+++ NEWS.md
@@ -7,6 +7,7 @@
 * Fix: `PipeOpImputeLearner` no longer adds `"factor"` or `"ordered"` levels for these feature types arbitrarily and instead updates levels correctly in certain edge-cases.
 * Fixed the error message for unexpected Multiplicities in the input and output type checking during `PipeOp`s training and prediction.
 * Untrained `PipeOp`s that take `NULL` as input during training now automatically perform training during prediction.
+* `PipeOpImputeConstant`, `PipeOpImputeMode`, `PipeOpImputeOOR`, and `PipeOpImputeLearner` can now handle `factor` or `ordered` features with zero levels.
 
 # mlr3pipelines 0.8.0